+++ src/Schema/Builder.php
@@ -2,12 +2,13 @@
 
 namespace StellarWP\Schema;
 
-use StellarWP\Schema\Container;
+use lucatume\DI52\App;
 use StellarWP\Schema\Fields;
+use StellarWP\Schema\Fields\Contracts\Schema_Interface as Field_Schema_Interface;
 use StellarWP\Schema\StellarWP\DB\DB;
 use StellarWP\Schema\StellarWP\DB\Database\Exceptions\DatabaseQueryException;
 use StellarWP\Schema\Tables;
-use StellarWP\Schema\Tables\Table_Schema_Interface;
+use StellarWP\Schema\Tables\Contracts\Schema_Interface as Table_Schema_Interface;
 use StellarWP\Schema\Tables\Filters\Group_FilterIterator;
 use WP_CLI;
 
@@ -25,7 +26,7 @@ class Builder {
 	 * @param Container $container Container instance.
 	 */
 	public function __construct( $container = null ) {
-		$this->container = $container ?: Container::init();
+		$this->container = $container ?: App::container();
 	}
 
 	/**
@@ -82,7 +83,7 @@ class Builder {
 		 *
 		 * @since 1.0.0
 		 *
-		 * @param \Iterator $table_schemas A list of Custom_Table_Interface objects that will have their tables dropped.
+		 * @param \Iterator $table_schemas A list of Table_Schema_Interface objects that will have their tables dropped.
 		 */
 		$table_schemas = apply_filters( 'stellarwp_tables_to_drop', $table_schemas );
 
@@ -111,7 +112,7 @@ class Builder {
 		 *
 		 * @since 1.0.0
 		 *
-		 * @param \Iterator $field_classes A list of Custom_Field_Interface objects that will have their fields dropped.
+		 * @param \Iterator $field_classes A list of Field_Schema_Interface objects that will have their fields dropped.
 		 */
 		$field_schemas = apply_filters( 'stellarwp_fields_to_drop', $field_schemas );