+++ src/Schema/Tables/Contracts/Table.php
@@ -1,13 +1,14 @@
 <?php
 
-namespace StellarWP\Schema\Tables;
+namespace StellarWP\Schema\Tables\Contracts;
 
-use StellarWP\Schema\Container;
+use lucatume\DI52\App;
+use lucatume\DI52\Container;
 use StellarWP\Schema\Fields;
-use StellarWP\Schema\Fields\Field_Schema_Interface;
+use StellarWP\Schema\Fields\Contracts\Schema_Interface as Field_Schema_Interface;
 use StellarWP\Schema\StellarWP\DB\DB;
 
-abstract class Abstract_Table implements Table_Schema_Interface {
+abstract class Table implements Schema_Interface {
 	/**
 	 * @var string|null The version number for this schema definition.
 	 */
@@ -62,7 +63,7 @@ abstract class Abstract_Table implements Table_Schema_Interface {
 	 * @param Container $container The container to use.
 	 */
 	public function __construct( $container = null ) {
-		$this->container = $container ?: Container::init();
+		$this->container = $container ?: App::container();
 	}
 
 	/**
@@ -154,7 +155,7 @@ abstract class Abstract_Table implements Table_Schema_Interface {
 		 *
 		 * @param string $base_table_name The base table name.
 		 * @param string $table_name The full table name.
-		 * @param Table_Schema_Interface $table_schema The table schema to be dropped.
+		 * @param Schema_Interface $table_schema The table schema to be dropped.
 		 */
 		do_action( 'stellarwp_pre_drop_table', $base_table_name, $this_table, $this );
 
@@ -175,7 +176,7 @@ abstract class Abstract_Table implements Table_Schema_Interface {
 		 *
 		 * @param string $base_table_name The base table name.
 		 * @param string $table_name The full table name.
-		 * @param Table_Schema_Interface $table_schema The table schema to be dropped.
+		 * @param Schema_Interface $table_schema The table schema to be dropped.
 		 */
 		do_action( 'stellarwp_post_drop_table', $base_table_name, $this_table, $this );
 
@@ -192,7 +193,7 @@ abstract class Abstract_Table implements Table_Schema_Interface {
 		 *
 		 * @param string $base_table_name The base table name.
 		 * @param string $table_name The full table name.
-		 * @param Table_Schema_Interface $table_schema The table schema to be dropped.
+		 * @param Schema_Interface $table_schema The table schema to be dropped.
 		 */
 		do_action( 'stellarwp_post_drop_table_wpdb_update', $base_table_name, $this_table, $this );
 
@@ -485,7 +486,7 @@ abstract class Abstract_Table implements Table_Schema_Interface {
 		 * @since 1.1.0
 		 *
 		 * @param string $table_name The prefix-less table name.
-		 * @param Abstract_Table $table The table object.
+		 * @param Table $table The table object.
 		 * @param \Iterator $field_schemas An iterable collection of field schemas associated with this table.
 		 */
 		do_action( 'stellarwp_schema_table_before_updete', static::table_name(), $this, $field_schemas );
@@ -502,7 +503,7 @@ abstract class Abstract_Table implements Table_Schema_Interface {
 		 *
 		 * @param string $table_name The prefix-less table name.
 		 * @param array $results The results of the schema update.
-		 * @param Abstract_Table $table The table object.
+		 * @param Table $table The table object.
 		 * @param \Iterator $field_schemas An iterable collection of field schemas associated with this table.
 		 */
 		do_action( 'stellarwp_schema_table_before_field_schema_updete', static::table_name(), $results, $this, $field_schemas );
@@ -518,7 +519,7 @@ abstract class Abstract_Table implements Table_Schema_Interface {
 		 *
 		 * @param string $table_name The prefix-less table name.
 		 * @param array $results The results of the table and field schema updates.
-		 * @param Abstract_Table $table The table object.
+		 * @param Table $table The table object.
 		 * @param \Iterator $field_schemas An iterable collection of field schemas associated with this table.
 		 */
 		do_action( 'stellarwp_schema_table_after_updete', static::table_name(), $results, $this, $field_schemas );