+++ tests/_support/Traits/Table_Fixtures.php
@@ -5,9 +5,8 @@ namespace StellarWP\Schema\Tests\Traits;
 use StellarWP\Schema\Activation;
 use StellarWP\Schema\Builder;
 use StellarWP\Schema\Container;
-use StellarWP\Schema\Fields\Abstract_Field;
-use StellarWP\Schema\Register;
-use StellarWP\Schema\Tables\Abstract_Table;
+use StellarWP\Schema\Fields\Contracts\Field;
+use StellarWP\Schema\Tables\Contracts\Table;
 
 trait Table_Fixtures {
 	private function assert_custom_tables_exist() {
@@ -50,7 +49,7 @@ trait Table_Fixtures {
 	 * Get a fake table to verify its creation.
 	 */
 	public function get_modified_simple_table() {
-		$table  = new class extends Abstract_Table {
+		$table  = new class extends Table {
 			const SCHEMA_VERSION = '2.0.0';
 			protected static $base_table_name = 'simple';
 			protected static $group = 'bork';
@@ -83,7 +82,7 @@ trait Table_Fixtures {
 	 * Get a fake table to verify its creation.
 	 */
 	public function get_simple_table() {
-		$table  = new class extends Abstract_Table {
+		$table  = new class extends Table {
 			const SCHEMA_VERSION = '1.0.0';
 
 			protected static $base_table_name = 'simple';
@@ -115,7 +114,7 @@ trait Table_Fixtures {
 	 * Get a fake table to verify its creation.
 	 */
 	public function get_indexless_table() {
-		$table  = new class extends Abstract_Table {
+		$table  = new class extends Table {
 			const SCHEMA_VERSION = '1.0.0';
 
 			protected static $base_table_name = 'noindex';
@@ -145,7 +144,7 @@ trait Table_Fixtures {
 	 * Get a simple table field class.
 	 */
 	public function get_simple_table_field() {
-		$field  = new class extends Abstract_Field {
+		$field  = new class extends Field {
 			const SCHEMA_VERSION = '1.0.0';
 
 			protected static $base_table_name = 'simple';