+++ tests/wpunit/BuilderTest.php
@@ -2,7 +2,7 @@
 
 namespace StellarWP\Schema\Tests\Builder;
 
-use StellarWP\Schema\Tables\Table_Schema_Interface;
+use StellarWP\Schema\Tables\Contracts\Schema_Interface as Table_Schema_Interface;
 use StellarWP\Schema\Register;
 use StellarWP\Schema\Schema;
 use StellarWP\Schema\Tests\SchemaTestCase;