+++ tests/wpunit/RegisterTest.php
@@ -2,11 +2,10 @@
 
 namespace StellarWP\Schema\Tests;
 
-use StellarWP\Schema\Container;
+use lucatume\DI52\App;
 use StellarWP\Schema\Fields;
 use StellarWP\Schema\Register;
 use StellarWP\Schema\Schema;
-use StellarWP\Schema\Tables;
 use StellarWP\Schema\Tests\Traits\Table_Fixtures;
 
 class RegisterTest extends SchemaTestCase {
@@ -22,7 +21,7 @@ class RegisterTest extends SchemaTestCase {
 
 		Register::field( $field_1 );
 
-		$this->assertArrayHasKey( $field_1::get_schema_slug(), Container::init()->make( Fields\Collection::class ) );
+		$this->assertArrayHasKey( $field_1::get_schema_slug(), App::make( Fields\Collection::class ) );
 	}
 
 	/**
@@ -37,7 +36,7 @@ class RegisterTest extends SchemaTestCase {
 			$field_1,
 		]);
 
-		$this->assertArrayHasKey( $field_1::get_schema_slug(), Container::init()->make( Fields\Collection::class ) );
+		$this->assertArrayHasKey( $field_1::get_schema_slug(), App::make( Fields\Collection::class ) );
 	}
 
 	/**